# zplus-api-develop/docs/swagger.yaml
definitions:
...
  dto.SelectedToolsDTO:
    properties:
      email:
        type: string
      tool_ids:
        items:
          type: string
        type: array
    required:
    - tool_ids
    type: object
...
  utils.ApiResponse:
    properties:
      data: {}
      error: {}
      message:
        type: string
    type: object
info:
  contact: {}
  description: API application for zplus security tool.
  title: Zplus Backend API
  version: "1.0"
paths:
...
  /select-tools:
    get:
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ApiResponse'
      security:
      - Bearer: []
      summary: Get the selected tools for the current session
      tags:
      - Tool Selection
    post:
      parameters:
      - description: ' '
        in: body
        name: select_tools
        required: true
        schema:
          $ref: '#/definitions/dto.SelectedToolsDTO'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ApiResponse'
      security:
      - Bearer: []
      summary: Select the tools for the current user session
      tags:
      - Tool Selection
  /tools:
    get:
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ApiResponse'
      security:
      - Bearer: []
      summary: Get all the enabled tools.
      tags:
      - Tool Selection
...
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"